Move the competence chart component
