+++ static/css/components/common/competences-chart.scss
